backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
openai==1.3.0
anthropic==0.7.7
requests==2.31.0
aiofiles==23.2.1
email-validator==2.1.0
jinja2==3.1.2
motor==3.3.2
beanie==1.23.6
httpx==0.25.2
celery==5.3.4
redis==5.0.1